+++ src/main/java/uk/ac/ebi/subs/data/component/Protocol.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.subs.data;
+package uk.ac.ebi.subs.data.component;
 
 public class Protocol {
     String name;
