+++ src/main/java/uk/ac/ebi/subs/data/component/SequencingSubmittableFile.java
@@ -1,4 +1,4 @@
-package uk.ac.ebi.subs.data;
+package uk.ac.ebi.subs.data.component;
 
 public class SequencingSubmittableFile extends SubmittableFile {
     public String readLabel;
