+++ src/main/java/uk/ac/ebi/subs/data/submittable/SampleGroup.java
@@ -1,9 +1,11 @@
-package uk.ac.ebi.subs.data;
+package uk.ac.ebi.subs.data.submittable;
+
+import uk.ac.ebi.subs.data.component.Attribute;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class SampleGroup  extends Submittable{
+public class SampleGroup  extends Submittable {
 
     List<String> sampleReferenceList = new ArrayList<String>();
