+++ src/main/java/uk/ac/ebi/subs/data/submittable/Study.java
@@ -1,4 +1,8 @@
-package uk.ac.ebi.subs.data;
+package uk.ac.ebi.subs.data.submittable;
+
+import uk.ac.ebi.subs.data.component.Publication;
+import uk.ac.ebi.subs.data.component.Attribute;
+import uk.ac.ebi.subs.data.component.Contact;
 
 import java.util.ArrayList;
 import java.util.Date;
